==> scripts/NM-recognition/gcms-processing/0_import_GCMSdata.R <==
# Install pacman package if needed ----
# install.packages('pacman')

# Load required packages ----
NPacks <- c("dplyr", "readr", "GCalignR", "stringr", "readxl", "here")
pacman::p_load(char = NPacks)

# FUNCTIONS ----
## A function to load CSV generated by MassHunterQual with lapply
my_read_csv  <- function(file, skip) {
  skip = 2
  read_csv(file = file, skip = skip)
}

# Import data files ----
## path to folder with GCMS-integration data files
#gcms_batch <- "0303-0903_2022"

## Make a list of the individual sample files paths
## Nu-IW ####
#path_samples_data
path_nu.iw_data <- list.files(path = here("data", "raw"
                                          , "Plasticity"
                                          , "gcms-integration", gcms_batch
                                          , "nu_iw"
                                          )
                                   , pattern = ".CSV" # Get all CSV files in the folder
                                   , full.names = T) %>% 
  str_subset('DR') # Do not include standards
#  str_subset('Hexane', negate = T)
#path_samples_data
path_nu.iw_data

## Import the files in the list, saving them into a list type object
nu.iw_data_list <- lapply(path_nu.iw_data, my_read_csv)
nu.iw_data_list %>% summary()

## PF-OW ####
#path_samples_data
path_pf.ow_data <- list.files(path = here("data", "raw"
                                          , "Plasticity"
                                          , "gcms-integration", gcms_batch
                                          , "pf_ow"
)
, pattern = ".CSV" # Get all CSV files in the folder
, full.names = T) %>% 
  str_subset('DR') # Do not include standards
#  str_subset('Hexane', negate = T)
#path_samples_data
path_pf.ow_data

## Import the files in the list, saving them into a list type object
pf.ow_data_list <- lapply(path_pf.ow_data, my_read_csv)
pf.ow_data_list %>% summary()

## Standards (STD) ####
path_standards <- list.files(path = here("data", "raw"
                                         , "Plasticity"
                                         , "gcms-integration"
                                         , gcms_batch)
                             , pattern = ".CSV" # Get all CSV files in the folder
                             , full.names = T) %>% 
  str_subset('STD') # Do not include samples
path_standards

## Import the files in the list, saving them into a list type object
standards_list <- lapply(path_standards, my_read_csv)
standards_list %>% summary()

# Rename the data frames in the list as the samples to which they belong ----
## Nu-IW ####
names(nu.iw_data_list) <- str_split(path_nu.iw_data
                                      , "/"
                                      , simplify = T) %>% 
  str_subset(".CSV") %>% 
  str_remove("DR-PL_") %>% 
  str_remove(".CSV") %>% 
  str_replace("-", "_")
names(nu.iw_data_list)

## PF-OW ####
names(pf.ow_data_list) <- str_split(path_pf.ow_data
                                    , "/"
                                    , simplify = T) %>% 
  str_subset(".CSV") %>% 
  str_remove("DR-PL_") %>% 
  str_remove(".CSV") %>% 
  str_replace("-", "_")
names(pf.ow_data_list)


## STD ####
names(standards_list) <- str_split(path_standards
                                   , "/"
                                   , simplify = T) %>% 
  str_subset(".CSV") %>% 
  str_remove("STD") %>% 
  str_remove("_SSL_") %>% 
  str_remove(".CSV") 
names(standards_list)

# Modify data frames to keep the necessary columns ----
## Nu-IW ####
names(nu.iw_data_list)
nu.iw_data_list %>% summary()
nu.iw_sample.names <- names(nu.iw_data_list)
#nu.iw_data_list[[eval(nu.iw_sample.names[1])]]
for (df in nu.iw_sample.names) {
  nu.iw_data_list[[df]] <- nu.iw_data_list[[df]] %>% 
    select("Center X", "Area") %>% 
    rename(RT = "Center X")
  #assign(df, df_n)
  #rm(df, df_n)
}
nu.iw_data_list %>% summary()

## PF-OW ####
names(pf.ow_data_list)
pf.ow_data_list %>% summary()
pf.ow_sample.names <- names(pf.ow_data_list)
#pf.ow_data_list[[eval(pf.ow_sample.names[1])]]
for (df in pf.ow_sample.names) {
  pf.ow_data_list[[df]] <- pf.ow_data_list[[df]] %>% 
    select("Center X", "Area") %>% 
    rename(RT = "Center X")
  #assign(df, df_n)
  #rm(df, df_n)
}
pf.ow_data_list %>% summary()

## STD ####
names(standards_list)
standards_list %>% summary()
standards.names <- names(standards_list)
#standards_list[[eval(OW_sample.names[1])]]
for (df in standards.names) {
  standards_list[[df]] <- standards_list[[df]] %>% 
    select("Center X", "Area") %>% 
    rename(RT = "Center X")
}
standards_list %>% summary()

# Input check-up for GCalignR ----
pdf(here("output"
         , "Plasticity"
         , paste0("GCalignR_input-check_PL_"
                  , gcms_batch
                  , '.pdf'))
    , width = 10, height = 5)

## Nu-IW ####
check_input(nu.iw_data_list, plot = T)
peak_interspace(nu.iw_data_list
                , rt_col_name = "RT"
                , quantile_range = c(0, 0.8)
                , quantiles = 0.05)

## PF-OW ####
check_input(pf.ow_data_list, plot = T)
peak_interspace(pf.ow_data_list
                , rt_col_name = "RT"
                , quantile_range = c(0, 0.8)
                , quantiles = 0.05)

## STD ####
check_input(standards_list, plot = T)
peak_interspace(standards_list
                , rt_col_name = "RT"
                , quantile_range = c(0, 0.8)
                , quantiles = 0.05)
#### Close graphic device to export plots into the PDF file
dev.off()
print("Input check-up plots were exported")

# Export data ----
save(list = c("nu.iw_data_list", "pf.ow_data_list", "standards_list")
     , file = here("data", "tmp"
                   ,"Plasticity", "data2align", paste0(gcms_batch
                                                       , ".Rdata")))

print("The data list(s) to align have been exported to the tmp data folder")

# End ----
## Report session information
capture.output(sessionInfo()
               , file = here("output", "Plasticity", "SInf_Script01.txt"))

# ## Detach/unload packages
# lapply(NPacks, unloadNamespace)
# 
# ## Clear environment
# rm(list=ls())
